// FE/medical-supplies-website/src/app/module/customer/model/CustomerUserDetail.ts
export interface CustomerUserDetail {
  customerId?: number;
  customerName?: string;
  customerCode?: string;
  phone?: string;
  gender?: boolean;
  customerAddress?: string;
  dateOfBirth?: string;
  idCard?: string;
  customerImg?: string;
  username?: string;
  accountEmail?: string;
  customerTypeName?: string;
}